# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "WALLET / КОШЕЛЕК API docs",
    "description": "API documentation for the final GoIT team project / Документация по API для финального командного проекта GoIT [Backend github repository](https://github.com/DenysPhV/project_group_2-back)",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "servers": [{ "url": "http://localhost:3001" }],
...
  "paths": {
    "/api/users/signup": {
      "post": {
        "tags": ["Users"],
        "summary": "User registration / Регистрация пользователя",
        "parameters": [],
        "requestBody": {
          "description": "Registration's object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                        "type": "object",
                          "required": ["name","email", "password"],
                           "properties": {
                             "name": {
                                "type": "string",
                                "description": "The name of the user",
                                "example": "Bob"
                              }, 
                             "email": {
                              "type": "string",
                              "description": "User's email",
                              "format": "email"
                              },
                             "password": {
                             "type": "string",
                             "description": "User's password",
                              "example": "qwerty123"
                             }
                            }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                           "properties": {
                             "user":{
                               "type":"object",
                               "properties":{
"name": {
                                    "type": "string",
                                    "description": "The name of the user",
                                    "example": "Bob"
                                  }, 
                                "email": {
                                  "type": "string",
                                  "description": "User's email",
                                  "format": "email"
                                  },
                                "id": {
                                "type": "string",
                                "description": "User's id",
                                "example": "507f1f77bcf86cd799439011"
                                  }
                               }
                             }
                                  
                            }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          },
          "409": {
            "description": "Provided email already exists",
            "content": {}
          }
        }
      }
    },
    "/api/users/login": {
    "post": {
        "tags": ["Users"],
        "summary": "User authentication / Аутентификация пользователя",
        "parameters": [],
        "requestBody": {
          "description": "Registration's object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                          "required": ["email", "password"],
                           "properties": {
                              "email": {
                              "type": "string",
                              "description": "User's email",
                              "format": "email"
                              },
                             "password": {
                             "type": "string",
                             "description": "User's password",
                              "example": "qwerty123"
                             }
                            }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type":"object",
                  "properties":{
                   "token": {
                        "type": "string",
                        "description": "token",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI1ZmMyMDg1YmQwOTM2NTI4MTA3Y2UyNzQiLCJzaWQiOiI1ZmMyZDJmY2UxZDIwNTA2NzAyYmRkMjIiLCJpYXQiOjE2MDY2MDM1MTYsImV4cCI6MTYwNjYwNzExNn0.rJ_QjU4KvA76H96RHsvOBChK0Vjbd0NmqjMxdQVJIXB"
                      },
                      "user": {
                        "type":"object",
                        "properties":{
                               "name": {
                                    "type": "string",
                                    "description": "The name of the user",
                                    "example": "Bob"
                                  }, 
                                "email": {
                                  "type": "string",
                                  "description": "User's email",
                                  "format": "email"
                                  }
                        }
                      }
                  
                  }
                  
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          },
          "401": {
            "description": "Email or password is wrong",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}